requirements.txt
pandas
matplotlib
numpy
openpyxl
pillow
ttkbootstrap
xlrd